pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "flask-marshmallow>=1.3.0",
    "marshmallow-sqlalchemy>=1.4.1",
    "flasgger>=0.9.7.1",
    "flask-cors>=5.0.1",
    "flask-jwt-extended>=4.7.1",
]
